pyproject.toml
[project]
name = "carvalima-helpdesk-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.122.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.4",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "python-jose>=3.5.0",
    "bcrypt>=5.0.0",
    "fastapi-mail>=1.5.8",
    "python-multipart>=0.0.20",
]
